smtpemail helper cleanup
