+++ test/helpers/smtpEmails.ts
@@ -15,7 +15,7 @@ const EmailItemSchema = z.object({
   isUnread: z.boolean(),
 })
 
-const EmailResponseSchema = z.object({
+export const EmailResponseSchema = z.object({
   results: z.array(EmailItemSchema),
 })
 
@@ -39,7 +39,7 @@ export type Email = {
  * @param search this can be TO or FROM or in subject email address
  * @returns validated Email
  */
-async function checkEmails(search: string): Promise<Email> {
+export async function checkEmails(search: string): Promise<Email> {
   return new Promise((resolve, reject) => {
     const options = {
       hostname: host,
@@ -72,7 +72,7 @@ async function checkEmails(search: string): Promise<Email> {
   })
 }
 
-async function extractPin(emailId: string): Promise<string | null> {
+export async function extractPin(emailId: string): Promise<string | null> {
   const apiUrl = `${smtp4devUrl}/api/Messages/${emailId}/part/2/content`
   // Fetch the raw email content
   const response = await fetch(apiUrl)
@@ -92,7 +92,7 @@ async function extractPin(emailId: string): Promise<string | null> {
   }
 }
 
-async function extractInvite(emailId: string): Promise<string | null> {
+export async function extractInvite(emailId: string): Promise<string | null> {
   const apiUrl = `${smtp4devUrl}/api/Messages/${emailId}/plaintext`
   const response = await fetch(apiUrl)
   if (!response.ok) {
@@ -110,7 +110,7 @@ async function extractInvite(emailId: string): Promise<string | null> {
   }
 }
 
-async function findNewAdminEmail(oldAdminEmailId: string): Promise<Email> {
+export async function findNewAdminEmail(oldAdminEmailId: string): Promise<Email> {
   return new Promise((resolve, reject) => {
     const options = {
       hostname: host,
@@ -155,7 +155,7 @@ async function findNewAdminEmail(oldAdminEmailId: string): Promise<Email> {
   })
 }
 
-function getHostPort(url: string): { host: string | null; port: string | null } {
+export function getHostPort(url: string): { host: string | null; port: string | null } {
   const indexOfDoubleSlash = url.indexOf('//')
   const hostAndPort = url.substring(indexOfDoubleSlash + 2)
   const hostPortArr = hostAndPort.split(':')
@@ -167,4 +167,26 @@ function getHostPort(url: string): { host: string | null; port: string | null } 
   return { host, port }
 }
 
-export { checkEmails, extractInvite, extractPin, findNewAdminEmail, getHostPort }
+export async function clearSmtp4devMessages() {
+  return new Promise((resolve, reject) => {
+    const options = {
+      hostname: 'localhost',
+      port: 5001,
+      path: '/api/messages/*', // Delete all messages
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    }
+
+    const req = http.request(options, (res) => {
+      if (res.statusCode === 200) {
+        resolve(true)
+      } else {
+        reject(new Error(`Failed to clear messages, status code: ${res.statusCode}`))
+      }
+    })
+    req.on('error', (error) => reject(error))
+    req.end()
+  })
+}
